Move test_at_most_one_current_script() to test_model_script.py
