+++ src/nextline_rdb/models/strategies/tests/test_st_script.py
@@ -1,14 +1,12 @@
-import pytest
 from hypothesis import given
 from hypothesis import strategies as st
 from sqlalchemy import select
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import selectinload
 
 from nextline_rdb.db import DB
 from nextline_rdb.utils.strategies import st_none_or
 
-from ... import CurrentScript, Model, Script
+from ... import Model, Script
 from .. import st_model_script
 
 
@@ -31,16 +29,3 @@ async def test_db(script: Script) -> None:
             script_ = await session.scalar(select_script)
 
     assert repr(script) == repr(script_)
-
-
-async def test_at_most_one_current_script() -> None:
-    async with DB(use_migration=False, model_base_class=Model) as db:
-        async with db.session() as session:
-            current_script = CurrentScript(script=Script(script=''))
-            session.add(current_script)
-            await session.commit()
-
-            current_script = CurrentScript(script=Script(script=''))
-            session.add(current_script)
-            with pytest.raises(IntegrityError):
-                await session.commit()
